requirements.txt
ipython
pdf2image
gTTS
python-pptx